+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.15] - 2025-02-06
+
 ### Added
 - Data Transfer Object classes generation from swagger spec
 - JsonApiResponseFactory::createResponse() `meta` and `links` arguments 
@@ -123,7 +125,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - FastRoute Dispatcher generation from swagger operationIds
 
-[Unreleased]: https://github.com/FreeElephants/json-api-php-toolkit/compare/0.0.14...HEAD
+[Unreleased]: https://github.com/FreeElephants/json-api-php-toolkit/compare/0.0.15...HEAD
+[0.0.15]: https://github.com/FreeElephants/json-api-php-toolkit/compare/0.0.14...0.0.15
 [0.0.14]: https://github.com/FreeElephants/json-api-php-toolkit/compare/0.0.13...0.0.14
 [0.0.13]: https://github.com/FreeElephants/json-api-php-toolkit/compare/0.0.12...0.0.13
 [0.0.12]: https://github.com/FreeElephants/json-api-php-toolkit/compare/0.0.11...0.0.12